% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_compose_map_reference_panel}
\alias{amrc_compose_map_reference_panel}
\title{Compose a Map and Reference Plot as a Manuscript-Style Figure Panel}
\usage{
amrc_compose_map_reference_panel(
  map_plot,
  reference_plot,
  widths = c(1, 1),
  collect_guides = TRUE,
  tag_levels = "A"
)
}
\arguments{
\item{map_plot}{A phenotype or external map plot.}

\item{reference_plot}{A reference-distance relationship plot.}

\item{widths}{Relative panel widths.}

\item{collect_guides}{Logical; collect shared legends.}

\item{tag_levels}{Tag style passed to \code{patchwork::plot_annotation()}.}
}
\value{
A patchwork-composed plot object.
}
\description{
Compose a Map and Reference Plot as a Manuscript-Style Figure Panel
}
