% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pbp_type.R
\name{amrc_default_sequence_exclusions}
\alias{amrc_default_sequence_exclusions}
\title{Default Sequence Exclusions for the Legacy S. pneumoniae Genotype Map}
\usage{
amrc_default_sequence_exclusions()
}
\value{
A character vector of isolate identifiers.
}
\description{
These are the isolates removed in the legacy genotype-map script because of
clear indel mutations.
}
