% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_plot_association_model_comparison}
\alias{amrc_plot_association_model_comparison}
\title{Plot an Association-Model Comparison}
\usage{
amrc_plot_association_model_comparison(
  comparison_table,
  mode = c("change_counts", "p_values", "effect_sizes")
)
}
\arguments{
\item{comparison_table}{Output from \code{\link[=amrc_compare_association_models]{amrc_compare_association_models()}}.}

\item{mode}{Plot type: \code{"change_counts"}, \code{"p_values"}, or
\code{"effect_sizes"}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot an Association-Model Comparison
}
