% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_plot_cluster_feature_shifts}
\alias{amrc_plot_cluster_feature_shifts}
\title{Plot Ranked Cluster-Difference Features}
\usage{
amrc_plot_cluster_feature_shifts(
  feature_summary,
  cluster_pair = NULL,
  top_n = 20,
  feature_col = "feature",
  shift_col = "max_state_frequency_shift"
)
}
\arguments{
\item{feature_summary}{Output from
\code{\link[=amrc_find_cluster_differentiating_features]{amrc_find_cluster_differentiating_features()}}.}

\item{cluster_pair}{Optional length-2 character vector selecting one cluster
pair.}

\item{top_n}{Number of features to plot.}

\item{feature_col}{Feature column.}

\item{shift_col}{Frequency-shift column.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Ranked Cluster-Difference Features
}
