% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_plot_distance_histogram}
\alias{amrc_plot_distance_histogram}
\title{Plot Cluster-Separation Histograms}
\usage{
amrc_plot_distance_histogram(
  hist_data,
  fill_values = c(`Intra-Group` = "blue", `Inter-Group` = "red"),
  binwidth = 1
)
}
\arguments{
\item{hist_data}{Data frame containing \code{DistanceType} and \code{Distance}.}

\item{fill_values}{Named fill vector.}

\item{binwidth}{Histogram bin width.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Cluster-Separation Histograms
}
