% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_plot_heritability_summary}
\alias{amrc_plot_heritability_summary}
\title{Plot Heritability Summaries}
\usage{
amrc_plot_heritability_summary(
  heritability_table,
  trait_col = NULL,
  value_col = NULL
)
}
\arguments{
\item{heritability_table}{Table containing one row per response.}

\item{trait_col}{Response/trait column.}

\item{value_col}{Heritability column.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Heritability Summaries
}
