% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_plot_one_vs_two_dimensional_projection}
\alias{amrc_plot_one_vs_two_dimensional_projection}
\title{Plot One- vs Two-Dimensional Procrustes Projection}
\usage{
amrc_plot_one_vs_two_dimensional_projection(projection_data)
}
\arguments{
\item{projection_data}{Output table from
\code{\link[=amrc_compare_one_and_two_dimensional_maps]{amrc_compare_one_and_two_dimensional_maps()}}$unique_phenotypes or
\verb{$comparison}.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot One- vs Two-Dimensional Procrustes Projection
}
