% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_plot_variance_decomposition}
\alias{amrc_plot_variance_decomposition}
\title{Plot Variance-Decomposition Summaries}
\usage{
amrc_plot_variance_decomposition(
  variance_table,
  trait_col = NULL,
  component_col = NULL,
  value_col = NULL
)
}
\arguments{
\item{variance_table}{Table containing response/component/proportion columns.}

\item{trait_col}{Response/trait column.}

\item{component_col}{Variance-component column.}

\item{value_col}{Numeric proportion column.}
}
\value{
A \code{ggplot} object.
}
\description{
Plot Variance-Decomposition Summaries
}
