% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mds.R
\name{amrc_rotate_configuration}
\alias{amrc_rotate_configuration}
\title{Rotate a Configuration Matrix}
\usage{
amrc_rotate_configuration(configuration, degrees)
}
\arguments{
\item{configuration}{Numeric matrix with one row per isolate and one column
per dimension.}

\item{degrees}{Rotation angle in degrees.}
}
\value{
A rotated numeric matrix.
}
\description{
Rotate a Configuration Matrix
}
