% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robustness.R
\name{amrc_run_mds_collection}
\alias{amrc_run_mds_collection}
\title{Run a Collection of MDS Fits}
\usage{
amrc_run_mds_collection(
  distance_list,
  weight_matrices = NULL,
  lab_ids = NULL,
  ndim = 2,
  type = "ratio",
  init = "torgerson",
  ...
)
}
\arguments{
\item{distance_list}{List of \code{dist} objects.}

\item{weight_matrices}{Optional list of weight matrices.}

\item{lab_ids}{Optional character vector of isolate identifiers.}

\item{ndim}{Number of dimensions.}

\item{type}{Transformation type.}

\item{init}{Initialisation strategy.}

\item{...}{Additional arguments passed to \code{\link[=amrc_compute_mds]{amrc_compute_mds()}}.}
}
\value{
A list of simplified MDS fit objects.
}
\description{
Run a Collection of MDS Fits
}
