% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{amrc_theme_cartography}
\alias{amrc_theme_cartography}
\title{Theme Used Throughout the Legacy Cartography Plots}
\usage{
amrc_theme_cartography()
}
\value{
A \code{ggplot2} theme object.
}
\description{
Theme Used Throughout the Legacy Cartography Plots
}
